+++ src/Orc.Automation.Recorder/Orc.Automation.Recorder.csproj
@@ -28,9 +28,5 @@
     <ProjectReference Include="..\Orc.Automation\Orc.Automation.csproj" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Services\Interfaces\" />
-  </ItemGroup>
-
   <Import Project="$(MSBuildProjectDirectory)\..\Directory.build.shared.explicit.props" Condition="Exists('$(MSBuildProjectDirectory)\..\Directory.build.shared.explicit.props')" />
 </Project>
