+++ src/Orc.Automation/Win32/Delegates/EnumWindowsCallBackDelegate.cs
@@ -1,0 +1,5 @@
+﻿namespace Orc.Automation.Win32;
+
+using System;
+
+internal delegate bool EnumWindowsCallBackDelegate(IntPtr hwnd, IntPtr lParam);
