+++ src/Orc.Automation/Win32/Enums/InputTypes.cs
@@ -1,0 +1,7 @@
+﻿namespace Orc.Automation.Win32;
+
+internal static class InputTypes 
+{
+    public const int Mouse = 0;
+    public const int Keyboard = 1;
+}
