+++ src/Orc.Automation/Win32/Enums/KeyEvents.cs
@@ -1,0 +1,8 @@
+﻿namespace Orc.Automation.Win32;
+
+internal static class KeyEvents
+{
+    public const int Extendedkey = 0x0001;
+    public const int Keyup = 0x0002;
+    public const int Scancode = 0x0008;
+}
