+++ src/Orc.Automation/Win32/Enums/MouseEvents.cs
@@ -1,0 +1,6 @@
+﻿namespace Orc.Automation.Win32;
+
+internal static class MouseEvents
+{
+    internal const int Virtualdesk = 0x4000;
+}
