+++ src/Orc.Automation/Win32/Enums/MouseInputTypes.cs
@@ -1,0 +1,8 @@
+﻿namespace Orc.Automation.Win32;
+
+internal static class MouseInputTypes
+{
+    internal const int XButton1 = 0x0001;
+    internal const int XButton2 = 0x0002;
+    internal const int WheelDelta = 120;
+}
