+++ src/Orc.Automation/Win32/Enums/SendMessageTimeoutFlags.cs
