+++ src/Orc.Automation/Win32/Enums/SendMouseInputFlags.cs
@@ -1,0 +1,19 @@
+﻿namespace Orc.Automation.Win32;
+
+using System;
+
+[Flags]
+public enum SendMouseInputFlags
+{
+    Move = 0x0001,
+    LeftDown = 0x0002,
+    LeftUp = 0x0004,
+    RightDown = 0x0008,
+    RightUp = 0x0010,
+    MiddleDown = 0x0020,
+    MiddleUp = 0x0040,
+    XDown = 0x0080,
+    XUp = 0x0100,
+    Wheel = 0x0800,
+    Absolute = 0x8000,
+};
