+++ src/Orc.Automation/Win32/Enums/VKeys.cs
@@ -1,0 +1,7 @@
+﻿namespace Orc.Automation.Win32;
+
+internal static class VKeys
+{
+    public const int ShiftMask = 0x0100;
+    public const int CharMask = 0x00FF;
+}
