+++ src/Orc.Automation/Win32/Enums/VirtualScreens.cs
@@ -1,0 +1,9 @@
+﻿namespace Orc.Automation.Win32;
+
+internal static class VirtualScreens
+{
+    public const int SmX = 76;
+    public const int SmY = 77;
+    public const int SmCx = 78;
+    public const int SmCy = 79;
+}
