+++ src/Orc.Automation/Win32/Enums/WindowMessages.cs
@@ -1,0 +1,6 @@
+﻿namespace Orc.Automation.Win32;
+
+public static class WindowMessages
+{
+    public const uint WM_NULL = 0x0000;
+}
