+++ src/Orc.Automation/Win32/Structs/Input.cs
@@ -1,0 +1,10 @@
+﻿namespace Orc.Automation.Win32;
+
+using System.Runtime.InteropServices;
+
+[StructLayout(LayoutKind.Sequential)]
+internal struct Input
+{
+    public int type;
+    public Inputunion union;
+};
