+++ src/Orc.Automation/Win32/Structs/Inputunion.cs
@@ -1,0 +1,12 @@
+﻿namespace Orc.Automation.Win32;
+
+using System.Runtime.InteropServices;
+
+[StructLayout(LayoutKind.Explicit)]
+internal struct Inputunion
+{
+    [FieldOffset(0)]
+    public MouseInputStruct mouseInput;
+    [FieldOffset(0)]
+    public Keybdinput keyboardInput;
+};
