+++ src/Orc.Automation/Win32/Structs/MouseInputStruct.cs
@@ -1,0 +1,15 @@
+﻿namespace Orc.Automation.Win32;
+
+using System;
+using System.Runtime.InteropServices;
+
+[StructLayout(LayoutKind.Sequential)]
+internal struct MouseInputStruct
+{
+    public int dx;
+    public int dy;
+    public int mouseData;
+    public int dwFlags;
+    public int time;
+    public IntPtr dwExtraInfo;
+};
