+++ src/Orc.Automation/Win32/Structs/POINT.cs
@@ -1,0 +1,22 @@
+﻿namespace Orc.Automation.Win32;
+
+using System;
+using System.Runtime.InteropServices;
+using System.Windows;
+
+/// <summary>
+/// Struct representing a point.
+/// </summary>
+[Serializable]
+[StructLayout(LayoutKind.Sequential)]
+// ReSharper disable once InconsistentNaming
+public struct POINT
+{
+    public int X;
+    public int Y;
+
+    public static implicit operator Point(POINT point)
+    {
+        return new Point(point.X, point.Y);
+    }
+}
